Move conflict error reporting to its own module
